Get rid of the Impls.
